+++ src/components/board/inform/InformDetail.vue
@@ -2,7 +2,7 @@
 import {onMounted, ref} from 'vue';
 import {useRoute, useRouter} from 'vue-router';
 import {delFetch, getFetch, putFetch} from "@/stores/apiClient.js";
-import {formatDate} from "../../../stores/util.js";
+import {formatDate} from "@/stores/util.js";
 
 const router = useRouter();
 const route = useRoute();
